requirements.txt
accelerate==0.31.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
boto3==1.37.33
botocore==1.37.38
braceexpand==0.1.7
certifi==2025.11.12
charset-normalizer==3.4.4
clean-fid==0.1.35
click==8.2.2
contourpy==1.3.3
cycler==0.12.1
datasets==2.20.0
deepspeed==0.15.1
diffusers==0.32.0
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
duckdb==1.1.2
einops==0.8.0
ffmpeg-python==0.2.0
filelock==3.20.0
fire==0.5.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.5.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
kiwisolver==1.4.9
lightning==2.4.0
lightning-utilities==0.15.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.5
ninja==1.13.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.6.0
omegaconf==2.3.0
openai==2.14.0
opencv-python==4.12.0.88
packaging==24.2
pandas==2.3.3
Pillow==9.5.0
platformdirs==4.5.1
propcache==0.4.1
protobuf==4.25.8
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyarrow-hotfix==0.7
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
s3transfer==0.11.5
safetensors==0.6.2
scenedetect==0.6.7.1
scipy==1.16.3
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.44.0
setproctitle==1.3.7
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.1
termcolor==3.2.0
tokenizers==0.22.1
torch==2.5.0
torchmetrics==1.8.2
torchvision==0.20.0
tqdm==4.67.1
transformers==4.57.1
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.16.6
webdataset==0.2.86
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.23.0
